+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.116.5
+
 - Add `/remote` command to display connection URL, password and setup guide. Password is no longer shown in logs or welcome message.
 - Improve server port binding for remote.